# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recruweb recruitment platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  candidate: "candidate",
  employer: "employer",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  candidate: "candidate",
  employer: "employer",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
  company?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type JobEmploymentType =
  (typeof JobEmploymentType)[keyof typeof JobEmploymentType];

export const JobEmploymentType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
  remote: "remote",
} as const;

export interface Job {
  id: string;
  title: string;
  company: string;
  location: string;
  category: string;
  employmentType: JobEmploymentType;
  description: string;
  requirements?: string;
  salaryRange?: string;
  skills?: string[];
  employerId: string;
  applicantCount?: number;
  isActive: boolean;
  createdAt?: string;
}

export interface JobsListResponse {
  jobs: Job[];
  total: number;
  page?: number;
  totalPages?: number;
}

export type CreateJobBodyEmploymentType =
  (typeof CreateJobBodyEmploymentType)[keyof typeof CreateJobBodyEmploymentType];

export const CreateJobBodyEmploymentType = {
  "full-time": "full-time",
  "part-time": "part-time",
  contract: "contract",
  internship: "internship",
  remote: "remote",
} as const;

export interface CreateJobBody {
  title: string;
  location: string;
  category: string;
  employmentType: CreateJobBodyEmploymentType;
  description: string;
  requirements?: string;
  salaryRange?: string;
  skills?: string[];
}

export type JobStatsByCategoryItem = {
  category: string;
  count: number;
};

export type JobStatsByTypeItem = {
  type: string;
  count: number;
};

export interface JobStats {
  totalJobs: number;
  byCategory: JobStatsByCategoryItem[];
  byType: JobStatsByTypeItem[];
  recentJobs: Job[];
}

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewed: "reviewed",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface CandidateProfile {
  id: string;
  userId: string;
  name: string;
  email: string;
  phone?: string;
  location?: string;
  currentTitle?: string;
  bio?: string;
  skills?: string[];
  education?: string;
  experience?: string;
  resumeUrl?: string;
  resumePublicId?: string;
  createdAt?: string;
}

export interface Application {
  id: string;
  jobId: string;
  candidateId: string;
  status: ApplicationStatus;
  coverLetter?: string;
  job?: Job;
  candidate?: CandidateProfile;
  createdAt?: string;
}

export interface ApplicationsListResponse {
  applications: Application[];
  total: number;
}

export interface ApplyBody {
  jobId: string;
  coverLetter?: string;
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  pending: "pending",
  reviewed: "reviewed",
  shortlisted: "shortlisted",
  rejected: "rejected",
  hired: "hired",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export interface UpdateCandidateBody {
  phone?: string;
  location?: string;
  currentTitle?: string;
  bio?: string;
  skills?: string[];
  education?: string;
  experience?: string;
}

export interface ResumeUploadResponse {
  resumeUrl: string;
  resumePublicId: string;
  message?: string;
}

export interface EmployerProfile {
  id: string;
  userId: string;
  name: string;
  email: string;
  company: string;
  industry?: string;
  companySize?: string;
  location?: string;
  website?: string;
  description?: string;
  createdAt?: string;
}

export interface UpdateEmployerBody {
  company?: string;
  industry?: string;
  companySize?: string;
  location?: string;
  website?: string;
  description?: string;
}

export interface EmployerDashboard {
  totalJobs: number;
  activeJobs: number;
  totalApplications: number;
  pendingApplications: number;
  shortlistedApplications: number;
  recentApplications: Application[];
}

export type GetJobsParams = {
  search?: string;
  location?: string;
  category?: string;
  employmentType?: string;
  page?: number;
  limit?: number;
};