# lib/api-zod/src/generated/types/employerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Recruweb recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployerProfile {
  id: string;
  userId: string;
  name: string;
  email: string;
  company: string;
  industry?: string;
  companySize?: string;
  location?: string;
  website?: string;
  description?: string;
  createdAt?: string;
}